+++ shared/src/commonMain/kotlin/at/asitplus/wallet/app/common/WalletConfig.kt
@@ -23,8 +23,9 @@ class WalletConfig(
 ) : SettingsRepository {
     private val config: Flow<ConfigData> =
         dataStoreService.getPreference(Configuration.DATASTORE_KEY_CONFIG).map {
-            it?.let { joseCompliantSerializer.decodeFromString<ConfigData>(it) }
-                ?: ConfigDataDefaults
+            it?.let {
+                joseCompliantSerializer.decodeFromString<ConfigData>(it)
+            } ?: ConfigData()
         }
 
     override val host: Flow<String> = config.map {
@@ -33,6 +34,7 @@ class WalletConfig(
     }
     override val clientId: Flow<String> = config.map { it.clientId }
 
+    override val walletProviderHost = config.map { it.walletProviderHost }
     override val isConditionsAccepted: Flow<Boolean> = config.map { it.isConditionsAccepted }
     override val presentmentUseNegotiatedHandover: Flow<Boolean> = config.map { it.presentmentUseNegotiatedHandover }
     override val presentmentBleCentralClientModeEnabled: Flow<Boolean> = config.map { it.presentmentBleCentralClientModeEnabled }
@@ -47,6 +49,7 @@ class WalletConfig(
     override fun set(
         host: String?,
         clientId: String?,
+        walletProviderHost: String?,
         isConditionsAccepted: Boolean?,
         presentmentUseNegotiatedHandover: Boolean?,
         presentmentBleCentralClientModeEnabled: Boolean?,
@@ -63,6 +66,7 @@ class WalletConfig(
             val newConfig = ConfigData(
                 host = host ?: this@WalletConfig.host.first(),
                 clientId = clientId ?: this@WalletConfig.clientId.first(),
+                walletProviderHost = walletProviderHost ?: this@WalletConfig.walletProviderHost.first(),
                 isConditionsAccepted = isConditionsAccepted
                     ?: this@WalletConfig.isConditionsAccepted.first(),
                 presentmentUseNegotiatedHandover = presentmentUseNegotiatedHandover
@@ -127,8 +131,9 @@ class WalletConfig(
  */
 @Serializable
 private data class ConfigData(
-    val host: String,
     val clientId: String = SettingsRepository.DEFAULT_CLIENT_ID,
+    val host: String = "https://wallet.a-sit.at/m7",
+    val walletProviderHost: String = "https://wallet-provider.a-sit.plus",
     val isConditionsAccepted: Boolean = false,
     val presentmentUseNegotiatedHandover: Boolean = true,
     val presentmentBleCentralClientModeEnabled: Boolean = true,
@@ -139,10 +144,4 @@ private data class ConfigData(
     val presentmentAllowMultipleRequests: Boolean = false,
     val readerAutomaticallySelectTransport: Boolean = true,
     val connectionTimeout: Duration = 15.seconds,
-)
-
-private val ConfigDataDefaults = ConfigData(
-    host = "https://wallet.a-sit.at/m7",
-    clientId = SettingsRepository.DEFAULT_CLIENT_ID,
-    isConditionsAccepted = false,
 )