+++ shared/src/commonMain/kotlin/ui/presentation/DCAPIPresentationGraphViewModel.kt
@@ -5,8 +5,9 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import androidx.navigation.toRoute
 import at.asitplus.catching
-import at.asitplus.dcapi.request.DCAPIWalletRequest
 import at.asitplus.dif.PresentationDefinition
+import at.asitplus.openid.RequestParametersFrom
+import at.asitplus.signum.indispensable.josef.io.joseCompliantSerializer
 import at.asitplus.valera.resources.Res
 import at.asitplus.valera.resources.biometric_authentication_prompt_for_data_transmission_consent_title
 import at.asitplus.wallet.app.common.WalletMain
@@ -14,7 +15,6 @@ import at.asitplus.wallet.app.common.toDifInputDescriptorList
 import at.asitplus.wallet.lib.agent.SubjectCredentialStore
 import at.asitplus.wallet.lib.data.CredentialPresentation
 import at.asitplus.wallet.lib.data.CredentialPresentationRequest
-import at.asitplus.signum.indispensable.josef.io.joseCompliantSerializer
 import at.asitplus.wallet.lib.ktor.openid.OpenId4VpWallet
 import at.asitplus.wallet.lib.openid.PresentationExchangeMatchingResult
 import kotlinx.coroutines.flow.MutableStateFlow
@@ -34,16 +34,17 @@ class DCAPIPresentationGraphViewModel(
     val apiRequestSerialized = route.apiRequestSerialized
 
     val dcApiWalletRequest = catching {
-        joseCompliantSerializer.decodeFromString<DCAPIWalletRequest.IsoMdoc>(apiRequestSerialized)
+        joseCompliantSerializer.decodeFromString<RequestParametersFrom.IsoMdocDcApi>(apiRequestSerialized)
     }
 
-    val selectionProvider = MutableStateFlow<UiState<Pair<DCAPIWalletRequest.IsoMdoc, CredentialSelectionProvider<SubjectCredentialStore.StoreEntry>>>>(
+    val selectionProvider = MutableStateFlow<UiState<Pair<RequestParametersFrom.IsoMdocDcApi, CredentialSelectionProvider<SubjectCredentialStore.StoreEntry>>>>(
         UiStateLoading
     ).apply {
         viewModelScope.launch {
             value = try {
                 val unwrappedDcApiWalletRequest = dcApiWalletRequest.getOrThrow()
-                val descriptors = unwrappedDcApiWalletRequest.isoMdocRequest.deviceRequest.docRequests.toDifInputDescriptorList()
+                val descriptors =
+                    unwrappedDcApiWalletRequest.parameters.isoMdocRequest.deviceRequest.docRequests.toDifInputDescriptorList()
                 val presentationRequest = CredentialPresentationRequest.PresentationExchangeRequest(
                     presentationDefinition = PresentationDefinition(
                         inputDescriptors = descriptors
@@ -107,7 +108,7 @@ class DCAPIPresentationGraphViewModel(
 
     private suspend fun finalizeAuthorization(
         credentialPresentation: CredentialPresentation,
-        request: DCAPIWalletRequest.IsoMdoc
+        request: RequestParametersFrom.IsoMdocDcApi
     ): OpenId4VpWallet.AuthenticationSuccess {
         walletMain.keyMaterial.promptText =
             getString(Res.string.biometric_authentication_prompt_for_data_transmission_consent_title)
@@ -119,7 +120,7 @@ class DCAPIPresentationGraphViewModel(
 
     private suspend fun finalizationMethod(
         credentialPresentation: CredentialPresentation,
-        request: DCAPIWalletRequest.IsoMdoc
+        request: RequestParametersFrom.IsoMdocDcApi
     ): OpenId4VpWallet.AuthenticationSuccess = walletMain.presentationService.finalizeDCAPIIsoMdocPresentation(
         credentialPresentation = when (credentialPresentation) {
             is CredentialPresentation.PresentationExchangePresentation -> credentialPresentation