+++ shared/src/commonMain/kotlin/ui/viewmodels/intents/DCAPIAuthorizationIntentViewModel.kt
@@ -1,12 +1,10 @@
 package ui.viewmodels.intents
 
-import at.asitplus.dcapi.request.DCAPIWalletRequest
+import at.asitplus.openid.RequestParametersFrom
 import at.asitplus.wallet.app.common.WalletMain
 import at.asitplus.wallet.app.common.domain.BuildAuthenticationConsentPageFromAuthenticationRequestDCAPIUseCase
-import at.asitplus.signum.indispensable.josef.io.joseCompliantSerializer
 import at.asitplus.wallet.lib.oidvci.OAuth2Exception
 import domain.BuildAuthenticationConsentPageFromAuthenticationRequest
-import domain.BuildAuthenticationConsentPageFromAuthenticationRequestUriUseCase
 import io.github.aakira.napier.Napier
 import kotlinx.coroutines.CoroutineExceptionHandler
 import kotlinx.coroutines.Dispatchers
@@ -38,11 +36,13 @@ class DCAPIAuthorizationIntentViewModel(
     }
 
     fun process() = walletMain.scope.launch(Dispatchers.Default + coroutineExceptionHandler) {
-        val dcApiRequest = walletMain.platformAdapter.getCurrentDCAPIData().getOrThrow()
+        val dcApiRequest = walletMain.platformAdapter.getCurrentDCAPIVerificationData().getOrThrow()
 
         val successRoute = when (dcApiRequest) {
-            is DCAPIWalletRequest.OpenId4Vp -> buildConsentPageFromRequest(dcApiRequest)
-            is DCAPIWalletRequest.IsoMdoc -> buildConsentPageFromDcApiRequest(dcApiRequest)
+            is RequestParametersFrom.OpenId4VpDcApiMultiSigned -> TODO()
+            is RequestParametersFrom.OpenId4VpDcApiSigned -> buildConsentPageFromRequest(dcApiRequest)
+            is RequestParametersFrom.OpenId4VpDcApiUnsigned -> buildConsentPageFromRequest(dcApiRequest)
+            is RequestParametersFrom.IsoMdocDcApi -> buildConsentPageFromDcApiRequest(dcApiRequest)
         }.getOrThrow()
 
         onSuccess(successRoute)